requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
py-zabbix==1.1.7
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
requests==2.28.1
rubrik-cdm==2.0.10
six==1.16.0
urllib3==1.26.12
